HelpWidget and preload augmentor move to Agent; dietary banner to Shifts (nobodies-collective/Humans#1091)
- New ChromeSlots.BodyEnd; both Shell layouts render the slot where they
  invoked HelpWidget by name. Agent contributes HelpWidget there, so the
  widget disappears with the section
- DietaryMissingBanner + its keys move to Humans.Shifts (call sites already
  invoke by name — unchanged)
- AgentPreloadAugmentor moves to Humans.Agent.Services.Preload, registered
  in Agent's Section; Web registration + orphaned using removed
- Moved resx keys (Help_*/Agent_ panel/Dietary*) SharedResource -> section
  resources across all 6 locales
