+++ src/Humans.Base/Resources/SharedResource.ca.resx
@@ -1484,35 +1484,8 @@
   <data name="Camp_Plural" xml:space="preserve">
     <value>Barris</value>
   </data>
-  <data name="Agent_InputPlaceholder" xml:space="preserve">
-    <value>Pregunta com funciona el sistema…</value>
-  </data>
-  <data name="Agent_PanelTitle" xml:space="preserve">
-    <value>Assistent</value>
-  </data>
-  <data name="Agent_TermsLink" xml:space="preserve">
-    <value>Condicions IA</value>
-  </data>
-  <data name="Agent_Send" xml:space="preserve">
-    <value>Envia</value>
-  </data>
-  <data name="Agent_ViewHistory" xml:space="preserve">
-    <value>Historial</value>
-  </data>
   <data name="Agent_WidgetButton" xml:space="preserve">
     <value>Obrir l'assistent</value>
-  </data>
-  <data name="Help_WidgetButton" xml:space="preserve">
-    <value>Ajuda</value>
-  </data>
-  <data name="Help_Menu_TalkWithAgent" xml:space="preserve">
-    <value>Parlar amb l'Assistent</value>
-  </data>
-  <data name="Help_Menu_CreateIssue" xml:space="preserve">
-    <value>Crear incidència</value>
-  </data>
-  <data name="Help_Agent_IssueProposed" xml:space="preserve">
-    <value>No puc respondre això jo mateix. He redactat una incidència per a l'equip — revisa-la i envia-la, si us plau.</value>
   </data>
   <data name="Nav_Agent" xml:space="preserve">
     <value>Assistent</value>
@@ -2248,9 +2221,6 @@
   <data name="Todo_DietaryMedical_Done" xml:space="preserve"><value>Thanks — we've got it</value></data>
   <data name="Todo_DietaryMedical_Action" xml:space="preserve"><value>Fill out</value></data>
   <data name="Todo_DietaryMedical_NoShift_Pending" xml:space="preserve"><value>Tell us about your food needs so we're ready when you sign up.</value></data>
-  <data name="DietaryMissingBanner_Title" xml:space="preserve"><value>We need your dietary info</value></data>
-  <data name="DietaryMissingBanner_Body" xml:space="preserve"><value>You're signed up for a long shift. The cantina can't plan your meals until you tell us your food preferences.</value></data>
-  <data name="DietaryMissingBanner_Cta" xml:space="preserve"><value>Tell us now</value></data>
   <data name="Profile_DietaryMedical_PageTitle" xml:space="preserve"><value>Dietary &amp; medical info</value></data>
   <data name="Profile_DietaryMedical_DietaryPreference_Label" xml:space="preserve"><value>Dietary preference</value></data>
   <data name="Profile_DietaryMedical_DietaryPreference_Omnivore" xml:space="preserve"><value>Omnivore</value></data>
