+++ src/Humans.Base/Resources/SharedResource.es.resx
@@ -1484,35 +1484,8 @@
   <data name="Camp_Plural" xml:space="preserve">
     <value>Barrios</value>
   </data>
-  <data name="Agent_InputPlaceholder" xml:space="preserve">
-    <value>Pregunta cómo funciona el sistema…</value>
-  </data>
-  <data name="Agent_PanelTitle" xml:space="preserve">
-    <value>Asistente</value>
-  </data>
-  <data name="Agent_TermsLink" xml:space="preserve">
-    <value>Términos IA</value>
-  </data>
-  <data name="Agent_Send" xml:space="preserve">
-    <value>Enviar</value>
-  </data>
-  <data name="Agent_ViewHistory" xml:space="preserve">
-    <value>Historial</value>
-  </data>
   <data name="Agent_WidgetButton" xml:space="preserve">
     <value>Abrir el asistente</value>
-  </data>
-  <data name="Help_WidgetButton" xml:space="preserve">
-    <value>Ayuda</value>
-  </data>
-  <data name="Help_Menu_TalkWithAgent" xml:space="preserve">
-    <value>Hablar con el Asistente</value>
-  </data>
-  <data name="Help_Menu_CreateIssue" xml:space="preserve">
-    <value>Crear incidencia</value>
-  </data>
-  <data name="Help_Agent_IssueProposed" xml:space="preserve">
-    <value>No puedo responder a esto por mí mismo. He redactado una incidencia para el equipo — revísala y envíala, por favor.</value>
   </data>
   <data name="Nav_Agent" xml:space="preserve">
     <value>Asistente</value>
@@ -2248,9 +2221,6 @@
   <data name="Todo_DietaryMedical_Done" xml:space="preserve"><value>Gracias — ya tenemos lo necesario</value></data>
   <data name="Todo_DietaryMedical_Action" xml:space="preserve"><value>Rellenar</value></data>
   <data name="Todo_DietaryMedical_NoShift_Pending" xml:space="preserve"><value>Cuéntanos tus necesidades alimentarias para estar listos cuando te apuntes.</value></data>
-  <data name="DietaryMissingBanner_Title" xml:space="preserve"><value>Necesitamos tus datos alimentarios</value></data>
-  <data name="DietaryMissingBanner_Body" xml:space="preserve"><value>Estás apuntada a un turno largo. La cantina no puede planificar tus comidas hasta que nos digas tus preferencias.</value></data>
-  <data name="DietaryMissingBanner_Cta" xml:space="preserve"><value>Decírnoslo ahora</value></data>
   <data name="Profile_DietaryMedical_PageTitle" xml:space="preserve"><value>Información alimentaria y médica</value></data>
   <data name="Profile_DietaryMedical_DietaryPreference_Label" xml:space="preserve"><value>Preferencia alimentaria</value></data>
   <data name="Profile_DietaryMedical_DietaryPreference_Omnivore" xml:space="preserve"><value>Omnívoro/a</value></data>
