+++ src/Humans.Base/Resources/SharedResource.resx
@@ -717,16 +717,7 @@
   <!-- Issue Comment Email -->
 
   <!-- Agent Widget -->
-  <data name="Agent_InputPlaceholder" xml:space="preserve"><value>Ask a question about how the system works…</value></data>
-  <data name="Agent_PanelTitle" xml:space="preserve"><value>Assistant</value></data>
-  <data name="Agent_TermsLink" xml:space="preserve"><value>AI Terms</value></data>
-  <data name="Agent_Send" xml:space="preserve"><value>Send</value></data>
-  <data name="Agent_ViewHistory" xml:space="preserve"><value>History</value></data>
   <data name="Agent_WidgetButton" xml:space="preserve"><value>Open the helper</value></data>
-  <data name="Help_WidgetButton" xml:space="preserve"><value>Help</value></data>
-  <data name="Help_Menu_TalkWithAgent" xml:space="preserve"><value>Talk with the Assistant</value></data>
-  <data name="Help_Menu_CreateIssue" xml:space="preserve"><value>Create issue</value></data>
-  <data name="Help_Agent_IssueProposed" xml:space="preserve"><value>I can't answer this myself. I've drafted an issue for the team — please review and submit it.</value></data>
   <data name="Nav_Agent" xml:space="preserve"><value>Assistant</value></data>
 
   <!-- Magic Link Auth -->
@@ -1033,9 +1024,6 @@
   <data name="Todo_DietaryMedical_Done" xml:space="preserve"><value>Thanks — we've got it</value></data>
   <data name="Todo_DietaryMedical_Action" xml:space="preserve"><value>Fill out</value></data>
   <data name="Todo_DietaryMedical_NoShift_Pending" xml:space="preserve"><value>Tell us about your food needs so we're ready when you sign up.</value></data>
-  <data name="DietaryMissingBanner_Title" xml:space="preserve"><value>We need your dietary info</value></data>
-  <data name="DietaryMissingBanner_Body" xml:space="preserve"><value>You're signed up for a long shift. The cantina can't plan your meals until you tell us your food preferences.</value></data>
-  <data name="DietaryMissingBanner_Cta" xml:space="preserve"><value>Tell us now</value></data>
   <data name="Profile_DietaryMedical_PageTitle" xml:space="preserve"><value>Dietary &amp; medical info</value></data>
   <data name="Profile_DietaryMedical_DietaryPreference_Label" xml:space="preserve"><value>Dietary preference</value></data>
   <data name="Profile_DietaryMedical_DietaryPreference_Omnivore" xml:space="preserve"><value>Omnivore</value></data>
