+++ src/Humans.Web/Views/Shared/_AdminLayout.cshtml
@@ -84,6 +84,6 @@
             crossorigin="anonymous"></script>
     <script src="~/js/site.js" asp-append-version="true"></script>
     @await RenderSectionAsync("Scripts", required: false)
-    @await Component.InvokeAsync("HelpWidget")
+    <vc:chrome-slot name="body-end" />
 </body>
 </html>
