+++ src/Sections/Humans.Agent/AgentResource.ca.resx
@@ -100,4 +100,31 @@
   <data name="Agent_HistoryTitle" xml:space="preserve">
     <value>Assistent — Les meves converses</value>
   </data>
+  <data name="Agent_InputPlaceholder" xml:space="preserve">
+    <value>Pregunta com funciona el sistema…</value>
+  </data>
+  <data name="Agent_PanelTitle" xml:space="preserve">
+    <value>Assistent</value>
+  </data>
+  <data name="Agent_TermsLink" xml:space="preserve">
+    <value>Condicions IA</value>
+  </data>
+  <data name="Agent_Send" xml:space="preserve">
+    <value>Envia</value>
+  </data>
+  <data name="Agent_ViewHistory" xml:space="preserve">
+    <value>Historial</value>
+  </data>
+  <data name="Help_WidgetButton" xml:space="preserve">
+    <value>Ajuda</value>
+  </data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve">
+    <value>Parlar amb l'Assistent</value>
+  </data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve">
+    <value>Crear incidència</value>
+  </data>
+  <data name="Help_Agent_IssueProposed" xml:space="preserve">
+    <value>No puc respondre això jo mateix. He redactat una incidència per a l'equip — revisa-la i envia-la, si us plau.</value>
+  </data>
 </root>
