+++ src/Sections/Humans.Agent/AgentResource.de.resx
@@ -100,4 +100,31 @@
   <data name="Agent_HistoryTitle" xml:space="preserve">
     <value>Assistent — Meine Gespräche</value>
   </data>
+  <data name="Agent_InputPlaceholder" xml:space="preserve">
+    <value>Frage, wie das System funktioniert…</value>
+  </data>
+  <data name="Agent_PanelTitle" xml:space="preserve">
+    <value>Assistent</value>
+  </data>
+  <data name="Agent_TermsLink" xml:space="preserve">
+    <value>KI-Bedingungen</value>
+  </data>
+  <data name="Agent_Send" xml:space="preserve">
+    <value>Senden</value>
+  </data>
+  <data name="Agent_ViewHistory" xml:space="preserve">
+    <value>Verlauf</value>
+  </data>
+  <data name="Help_WidgetButton" xml:space="preserve">
+    <value>Hilfe</value>
+  </data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve">
+    <value>Mit dem Assistenten sprechen</value>
+  </data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve">
+    <value>Problem melden</value>
+  </data>
+  <data name="Help_Agent_IssueProposed" xml:space="preserve">
+    <value>Das kann ich selbst nicht beantworten. Ich habe einen Vorgang für das Team entworfen — bitte prüfe und sende ihn ab.</value>
+  </data>
 </root>
