+++ src/Sections/Humans.Agent/AgentResource.es.resx
@@ -100,4 +100,31 @@
   <data name="Agent_HistoryTitle" xml:space="preserve">
     <value>Asistente — Mis conversaciones</value>
   </data>
+  <data name="Agent_InputPlaceholder" xml:space="preserve">
+    <value>Pregunta cómo funciona el sistema…</value>
+  </data>
+  <data name="Agent_PanelTitle" xml:space="preserve">
+    <value>Asistente</value>
+  </data>
+  <data name="Agent_TermsLink" xml:space="preserve">
+    <value>Términos IA</value>
+  </data>
+  <data name="Agent_Send" xml:space="preserve">
+    <value>Enviar</value>
+  </data>
+  <data name="Agent_ViewHistory" xml:space="preserve">
+    <value>Historial</value>
+  </data>
+  <data name="Help_WidgetButton" xml:space="preserve">
+    <value>Ayuda</value>
+  </data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve">
+    <value>Hablar con el Asistente</value>
+  </data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve">
+    <value>Crear incidencia</value>
+  </data>
+  <data name="Help_Agent_IssueProposed" xml:space="preserve">
+    <value>No puedo responder a esto por mí mismo. He redactado una incidencia para el equipo — revísala y envíala, por favor.</value>
+  </data>
 </root>
