+++ src/Sections/Humans.Agent/AgentResource.fr.resx
@@ -100,4 +100,31 @@
   <data name="Agent_HistoryTitle" xml:space="preserve">
     <value>Assistant — Mes conversations</value>
   </data>
+  <data name="Agent_InputPlaceholder" xml:space="preserve">
+    <value>Posez une question sur le fonctionnement…</value>
+  </data>
+  <data name="Agent_PanelTitle" xml:space="preserve">
+    <value>Assistant</value>
+  </data>
+  <data name="Agent_TermsLink" xml:space="preserve">
+    <value>Conditions IA</value>
+  </data>
+  <data name="Agent_Send" xml:space="preserve">
+    <value>Envoyer</value>
+  </data>
+  <data name="Agent_ViewHistory" xml:space="preserve">
+    <value>Historique</value>
+  </data>
+  <data name="Help_WidgetButton" xml:space="preserve">
+    <value>Aide</value>
+  </data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve">
+    <value>Parler avec l'Assistant</value>
+  </data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve">
+    <value>Créer un signalement</value>
+  </data>
+  <data name="Help_Agent_IssueProposed" xml:space="preserve">
+    <value>Je ne peux pas répondre à cela moi-même. J'ai rédigé un signalement pour l'équipe — relis-le et envoie-le, s'il te plaît.</value>
+  </data>
 </root>
