+++ src/Sections/Humans.Agent/AgentResource.it.resx
@@ -100,4 +100,31 @@
   <data name="Agent_HistoryTitle" xml:space="preserve">
     <value>Assistente — Le mie conversazioni</value>
   </data>
+  <data name="Agent_InputPlaceholder" xml:space="preserve">
+    <value>Chiedi come funziona il sistema…</value>
+  </data>
+  <data name="Agent_PanelTitle" xml:space="preserve">
+    <value>Assistente</value>
+  </data>
+  <data name="Agent_TermsLink" xml:space="preserve">
+    <value>Condizioni IA</value>
+  </data>
+  <data name="Agent_Send" xml:space="preserve">
+    <value>Invia</value>
+  </data>
+  <data name="Agent_ViewHistory" xml:space="preserve">
+    <value>Cronologia</value>
+  </data>
+  <data name="Help_WidgetButton" xml:space="preserve">
+    <value>Aiuto</value>
+  </data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve">
+    <value>Parla con l'Assistente</value>
+  </data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve">
+    <value>Crea segnalazione</value>
+  </data>
+  <data name="Help_Agent_IssueProposed" xml:space="preserve">
+    <value>Non posso rispondere a questo da solo. Ho preparato una segnalazione per il team — controllala e inviala, per favore.</value>
+  </data>
 </root>
