+++ src/Sections/Humans.Agent/AgentResource.resx
@@ -72,4 +72,13 @@
   <data name="Agent_HistoryMessages" xml:space="preserve"><value>Messages</value></data>
   <data name="Agent_HistoryStarted" xml:space="preserve"><value>Started</value></data>
   <data name="Agent_HistoryTitle" xml:space="preserve"><value>Agent — My Conversations</value></data>
+  <data name="Agent_InputPlaceholder" xml:space="preserve"><value>Ask a question about how the system works…</value></data>
+  <data name="Agent_PanelTitle" xml:space="preserve"><value>Assistant</value></data>
+  <data name="Agent_TermsLink" xml:space="preserve"><value>AI Terms</value></data>
+  <data name="Agent_Send" xml:space="preserve"><value>Send</value></data>
+  <data name="Agent_ViewHistory" xml:space="preserve"><value>History</value></data>
+  <data name="Help_WidgetButton" xml:space="preserve"><value>Help</value></data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve"><value>Talk with the Assistant</value></data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve"><value>Create issue</value></data>
+  <data name="Help_Agent_IssueProposed" xml:space="preserve"><value>I can't answer this myself. I've drafted an issue for the team — please review and submit it.</value></data>
 </root>
