+++ src/Sections/Humans.Agent/Docs/Agent.md
@@ -129,7 +129,7 @@ Missing or wrong key → 401 (503 if the key is not configured). Unknown id → 
 
 ## Architecture
 
-**Owning services:** `AgentService` (orchestrator), `AgentSettingsService`, `AgentToolDispatcher`, `AgentUserSnapshotProvider`, `AgentAbuseDetector`, `AgentPromptAssembler`, `AgentPreloadCorpusBuilder`, `AnthropicClient`. `AgentPreloadAugmentor` stays in Shell (it reads Shell-owned help content) and implements `IAgentPreloadAugmentor` from this project's `Contracts/` folder; `AgentConversationRetentionJob` moved into the same folder at G5 lane 5b-5 and calls `IAgentConversationRetention` beside it. Only the job's DI registration and its roll-call entry stay in Shell.
+**Owning services:** `AgentService` (orchestrator), `AgentSettingsService`, `AgentToolDispatcher`, `AgentUserSnapshotProvider`, `AgentAbuseDetector`, `AgentPromptAssembler`, `AgentPreloadCorpusBuilder`, `AgentPreloadAugmentor`, `AnthropicClient`. `AgentPreloadAugmentor` moved into this project at nobodies-collective/Humans#1091 (it reads `Humans.Base`'s shared help registries, not anything Web-owned) and implements `IAgentPreloadAugmentor` from this project's `Contracts/` folder; `AgentConversationRetentionJob` moved into the same folder at G5 lane 5b-5 and calls `IAgentConversationRetention` beside it. Only the job's DI registration and its roll-call entry stay in Shell.
 **Owned tables:** `agent_conversations`, `agent_messages`, `agent_settings`.
 **Status:** (A) G5 — the section lives in its own project, `src/Sections/Humans.Agent`, with its cross-section surface in the project's own `Contracts/` folder (nobodies-collective/Humans#866, wave A4b); the `Humans.Agent.Contracts` leaf it once occupied folded back in (ruling 44). Everything except `Section`, `AgentResource`, the generated migrations, and the four `Contracts/` types (`AgentSectionKeys`, `IAgentAvailability`, `IAgentConversationRetention`, `IAgentPreloadAugmentor`) is `internal`, enforced at build time by HUM0034. Architecture tests: `tests/Humans.Agent.Tests/AgentArchitectureTests.cs`; page rendering: `tests/Humans.Integration.Tests/Controllers/AgentPageRenderTests.cs`. **No cross-section FK or nav at the EF level** — `agent_conversations.UserId`, `agent_messages.HandedOffToFeedbackId`, and `feedback_reports.AgentConversationId` are bare Guid columns.
