+++ src/Sections/Humans.Agent/Docs/data-access.md
@@ -13,12 +13,12 @@ Singleton like most other sections' repositories). Owns
 The preload/warmup surface: `AgentPreloadCorpusBuilder` assembles the
 tool-corpus preload from `AgentSectionDocReader`, `AgentFeatureSpecReader`,
 `CommunityFaqReader`, and calls into `IAgentPreloadAugmentor`
-(`Humans.Web.Services.Agent.AgentPreloadAugmentor`,
-`src/Humans.Web/Services/Agent/`) — a Web-layer helper, not part of this
-section, because it renders the access matrix / glossaries / route map /
-FAQ preload pages from `Humans.UI`'s `AccessMatrixDefinitions` /
-`SectionHelpContent` (every section's help content, which only the Web
-layer can see in one place). Pure static-content formatting — no DI
+(`Humans.Agent.Services.Preload.AgentPreloadAugmentor`,
+`src/Sections/Humans.Agent/Services/Preload/`, moved in at
+nobodies-collective/Humans#1091) — it renders the access matrix / glossaries /
+route map / FAQ preload pages from `Humans.Base`'s `AccessMatrixDefinitions` /
+`SectionHelpContent` (every section's help content, visible from any section
+since both live in the shared base layer). Pure static-content formatting — no DI
 dependencies beyond the two static readers, no DB access, no cache.
 `AgentPreloadWarmupHostedService` and `AgentSettingsStoreWarmupHostedService`
 run startup warmup, no DB access — fan out over the readers /
