+++ src/Sections/Humans.Agent/Views/Shared/Components/HelpWidget/Default.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.ViewComponents.HelpWidgetModel
+@model Humans.Agent.ViewComponents.HelpWidgetModel
 
 <!-- Floating "Help" bubble + popover menu -->
 <div class="help-widget position-fixed" style="bottom: 24px; right: 24px; z-index: 1050;">
