+++ src/Sections/Humans.Debug/Views/WidgetGallery/Index.cshtml
@@ -609,7 +609,7 @@
                             <span class="wg-card-name">&lt;vc:dietary-missing-banner&gt;</span>
                             <span class="wg-card-kind kind-vc">ViewComponent</span>
                         </div>
-                        <span class="wg-card-path">src/Humans.Web/ViewComponents/DietaryMissingBannerViewComponent.cs</span>
+                        <span class="wg-card-path">src/Sections/Humans.Shifts/ViewComponents/DietaryMissingBannerViewComponent.cs</span>
                     </div>
                     <div class="wg-card-note">
                         Red dietary-info nudge shown on /Shifts and /Shifts/Mine. Self-gates: renders only when the user has a
@@ -810,7 +810,7 @@
                             <span class="wg-card-name">&lt;vc:help-widget&gt;</span>
                             <span class="wg-card-kind kind-vc">ViewComponent</span>
                         </div>
-                        <span class="wg-card-path">src/Humans.Web/ViewComponents/HelpWidgetViewComponent.cs</span>
+                        <span class="wg-card-path">src/Sections/Humans.Agent/ViewComponents/HelpWidgetViewComponent.cs</span>
                     </div>
                     <div class="wg-card-note">
                         Floating "Help" launcher with a popover menu — replaces the previous separate Issues and Agent FABs.
